==> Supplements_Yasser_24/01_intro_pandas.ipynb <==
## DataFrames in Pandas

DataFrames are based on the idea of tables. As in SQL, a table consists of columns which describe attributes of the data. Each entry is a row in the table and fills each respective column with a value. The data type of the value in each column must be defined, however each column can hold values of different data types. 

## Importing data into pandas DataFrames

import pandas as pd
import numpy as np

Creating a DataFrame can be realised in various manners, e.g. here is an example of creating a DataFrame from a dictionary:

#create a dataframe
raw_data = {'medium': ['NYTimes', 'LATimes', 'WashingtonPost', 'BBC', 'TheGuardian'], 
        'location': ['New York', 'Los Angeles', "Washington DC", 'London', 'London'], 
        'articles': [500, '...', 815, 90, 390], 
        'sections': [19, 14, 15, '...', 21],
        'mainSection': ['world', 'politics', 'tech', 'u.s.', 'economy']}

df = pd.DataFrame(raw_data, columns = ['medium', 'location', 'articles', 'sections', 'mainSection'])
df

type(df)

Pandas provides an own data rype for DataFrames.

print(df.columns)

df.columns.values

res = df.medium
print(type(res))
print(res)

#convert it to a .csv file
df.to_csv('test.csv')

#read from a .csv file
df_new = pd.read_csv('test.csv', encoding='utf-8')

#display the data
df_new.head()

#load csv without a header
df_noheader = pd.read_csv('test.csv', header=None)
df_noheader.head()

#load csv wand specify column names
df_columns = pd.read_csv('test.csv', names = ['ID', 'medium', 'location', 'articles', 'sections', 'mainSection'])
df_columns.head()

df = pd.read_csv('test.csv', index_col='ID', names=['ID', 'medium', 'location', 'articles', 'sections', 'mainSection'])
df

df = pd.read_csv('test.csv', na_values=['...'])
df

pd.isnull(df)

#skip a specific number of rows
df = pd.read_csv('test.csv', skiprows=2)
df

#select a specific number of columns
df = pd.read_csv('test.csv', usecols=[1,3])
df

#export to and import from Excel
#note: you may need to install the packages xlwt and xlrd
#pip install xlwt
#pip install xlrd
df.to_excel('test.xls')

df_xls = pd.read_excel('test.xls')
df_xls

### Importing data from SQL tables

import sqlite3 
con = sqlite3.connect("data/guard_data.db")

dfsql = pd.read_sql_query("SELECT * FROM article_content",con)

dfsql

dfsql.to_csv('data/media.csv')

We can rename / reset the index of a DataFrame - let's say we want the values to descend from 240 to 0:

l = [x for x in range(0,240)][::-1]
dfsql.index = l
dfsql

Renaming the columns works just as well:

dfsql.columns = ['ID', 'content', 'comments', 'date']
dfsql

In our case, we imported some ID values from the SQL table already, hence we actually do not need the ID column. 

dfsql.index = dfsql.ID
dfsql

We now have no use for the 2ns ID column any more, let's remove it from the DataFrame: 

dfsql.drop('ID', axis = 1, inplace = True)
dfsql

__axis = 1__ tells us to remove a column, setting the __inplace__ parameter to __True__ tells Pandas to remove it directly in the DataFrame. 

### Useful functions in the DataFrame

print(dfsql.comments.sum())

print(dfsql.comments.cumsum())

dfsql.dtypes

dfsql.describe()

dfsql.isnull().sum()

dfsql['comments'].value_counts()

dfsql.head()

dfsql.head(10)

### Manipulating DataFrames

# adding a column
dfsql['new column'] = [x for x in range(0,dfsql.shape[0])]
dfsql

dfsql = dfsql.sort_values(by="comments", ascending=False)
dfsql.head()

dfsql['empty'] = np.nan
dfsql.head()

If you have only a few specific values for a column, you can assign them using the index of the DataFrame. In our case, we use the __'ID'__ column as index and fill the newly created empty column.

values = pd.Series([75, 70, 65], index=[1082, 1226, 1089])
dfsql['empty'] = values
dfsql
